[spirv-out] Move instructions to own module
