+++ src/back/spv/mod.rs
@@ -1,4 +1,5 @@
 mod helpers;
+mod instructions;
 mod layout;
 mod writer;
